pyproject.toml
[project]
name = "auto-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=46.0.5",
    "fastapi>=0.134.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "rich>=14.3.3",
]
